+++ tests/test_pipenv_pipes.py
@@ -6,9 +6,8 @@
 import pytest
 
 from click.testing import CliRunner
-
-from pipenv_pipes import pipenv_pipes
 from pipenv_pipes import cli
+from pipenv_pipes import utils
 
 
 @pytest.fixture
@@ -19,9 +18,9 @@ def _fixture():
 def test_command_line_interface():
     """Test the CLI."""
     runner = CliRunner()
-    result = runner.invoke(cli.main)
+    result = runner.invoke(cli.pipes)
     assert result.exit_code == 0
-    assert 'pipenv_pipes.cli.main' in result.output
-    help_result = runner.invoke(cli.main, ['--help'])
+    assert 'Pipenv Projects' in result.output
+    help_result = runner.invoke(cli.pipes, ['--help'])
     assert help_result.exit_code == 0
     assert '--help  Show this message and exit.' in help_result.output